requirements.txt
python-arango==7.5.2
#web3==5.31.1
click==8.0.1
requests
multithread-parallel-processing==1.0.0
pydantic~=1.10.4
psycopg2-binary
pymongo==3.11.4
python-dotenv==0.19.2
python-cli-scheduler
#websockets~=10.4
QueryStateLib==1.1.4
selenium==4.6.0
SQLAlchemy~=1.4.46
timeout-decorator

sanic==22.12.0
Sanic-Cors==2.2.0
sanic-ext==22.12.0
sanic-routing==22.8.0
sanic-testing==23.3.0
termcolor==2.2.0